% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rut.R
\name{as_rut}
\alias{as_rut}
\title{Formatear una cadena de texto o numero en formato RUT}
\usage{
as_rut(rut)
}
\arguments{
\item{rut}{RUT con o sin digito verificador, con o sin guion y/o puntos.

ruttexto_rutnum("16.019.432")}
}
\description{
Formatear una cadena de texto o numero en formato RUT
}
