% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idlatr-package.R
\name{clrutr-exports}
\alias{clrutr-exports}
\alias{\%>\%}
\title{idlatr exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the clrutr package to avoid listing the magrittr
as Depends of clrutr
}
