% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clrutr-package.R
\docType{package}
\name{clrutr-package}
\alias{clrutr}
\alias{clrutr-package}
\title{clrutr: Funciones para el tratamiento y formato de identificadores tributarios}
\description{
Funciones para el tratamiento y formato de identificadores tributarios.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jkunst.com/clrutr/}
}

}
\author{
\strong{Maintainer}: Joshua Kunst \email{jbkunst@gmail.com}

Authors:
\itemize{
  \item Joshua Kunst \email{jbkunst@gmail.com}
}

}
\keyword{internal}
