% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clrutr-package.R
\docType{package}
\name{clrutr}
\alias{clrutr}
\title{Funciones para el tratamiento y formato de identificadores tributarios}
\description{
Funciones para el tratamiento y formato de 
  identificadores tributarios
}
\author{
Joshua Kunst (@jbkunst)
}
