% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idlatr-package.R
\name{idlatr-exports}
\alias{\%>\%}
\alias{idlatr-exports}
\title{idlatr exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the idlatr package to avoid listing the magrittr
as Depends of idlatr.
}

