% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rut.R
\name{rut_rut10}
\alias{rut_rut10}
\title{Formatear RUT con digito verificador a RUT de largo 10}
\usage{
rut_rut10(rut)
}
\arguments{
\item{rut}{RUT con digito verificador.}
}
\description{
Formatear RUT con digito verificador a RUT de largo 10
}

