% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rut.R
\name{rutnum_dv}
\alias{rutnum_dv}
\title{Calculo digito verificador de RUTs}
\usage{
rutnum_dv(rutnum)
}
\arguments{
\item{rutnum}{ruts en formato numerico.}
}
\description{
Calculo digito verificador de RUTs
}
\examples{

rutnum <- c(16019432, 16355485, 15724861, 121, 6505922)
rutnum_dv(rutnum)

}

