% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rut.R
\name{rutnum_rut}
\alias{rutnum_rut}
\title{Formatear un rut numerico a rut con digito verificador}
\usage{
rutnum_rut(rutnum, sep = "")
}
\arguments{
\item{rutnum}{Rut en formato numerico}

\item{sep}{separador}
}
\description{
Formatear un rut numerico a rut con digito verificador
}
\examples{

rutnum <- c(16019432, 16355485, 15724861, 121, 6505922)

rutnum_rut(rutnum)
rutnum_rut(rutnum, sep = "-")

}

