% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rut.R
\name{rut10_rutnum}
\alias{rut10_rutnum}
\title{Formatear un RUT de largo 10 a RUT numerico}
\usage{
rut10_rutnum(rut10)
}
\arguments{
\item{rut10}{RUT con digito verificador de largo 10.

r10 <- rut_rut10(rutnum_rut(c(16019432, 16355485, 15724861, 121, 6505922)))

r10

rut10_rutnum(r10)}
}
\description{
Formatear un RUT de largo 10 a RUT numerico
}
