% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rut.R
\name{rut_rut10}
\alias{rut_rut10}
\title{Formatear rut con digito verificador a rut de largo 10}
\usage{
rut_rut10(rut)
}
\arguments{
\item{rut}{Rut con digito verificador.}
}
\description{
Formatear rut con digito verificador a rut de largo 10
}

