% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rut.R
\name{rutnum_rut}
\alias{rutnum_rut}
\title{Formatear RUT numerico a RUT con digito verificador}
\usage{
rutnum_rut(rutnum, sep = "")
}
\arguments{
\item{rutnum}{RUT en formato numerico.}

\item{sep}{Separador.}
}
\description{
Formatear RUT numerico a RUT con digito verificador
}

